+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Added `MeshSprite` which is used for rendering 3D objects as 2D sprites.
+
 ## [0.9.6] - 2021-07-28
 ### Added
 - Added `destroy` function to an instanced model.